Store descriptive metrics identified by pipeline run id
